// Frontend/src/app/models/segmentmeasures.ts
export interface SegmentMeasurements{
    id_medicion?: number;
    longitud_mano: number;
    longitud_palma: number;
    amplitud_mano_metacarpios: number;
    longitud_dedo_indice: number;
    anchura_dedo_indice_proximal: number;
    ancho_dedo_indice_distal: number;
    longitud_pie: number;
    ancho_pie: number;
    longitud_cabeza: number;
    ancho_cabeza: number;
    circunferencia_cabeza: number;
    arco_bitragion: number;
    longitud_pulgar: number;
    ancho_pulgar: number;
    circunferencia_brazo_flexionado: number;
    circunferencia_antebrazo_flexionado: number;
    id_trabajador_pertenece: number;
}