// Frontend/src/app/models/functionmeasures.ts
export interface FunctionalMeasurements{
  id_medicion?: number;
  alcance_agarre: number;
  longitud_codo_munieca: number;
  longitud_agarre_codo: number;
  altura_punio: number;
  profundidad_asiento: number;
  longitud_rodilla_gluteo: number;
  circunferencia_cintura: number;
  circunferencia_munieca: number;
  id_trabajador_pertenece: number;
}